Move TestBarycenter from test_barycenter to test_distance_measures
I don't know the nose package, so I was afraid of trying to develop the tests in
the existing test_distance_measures module. However nose seems to play nicely
with the standard unittest package, so we're all good.
